+++ bin/console
@@ -1,11 +1,13 @@
 #!/usr/bin/env ruby
+# frozen_string_literal: true
 
 require "bundler/setup"
 require "sequel"
 require "umbrellio-sequel-plugins"
-require "irb"
-require_relative "../database"
+require "pry"
+
+require_relative "../utils/database"
 
 Dir["#{__dir__}/lib/**/*.rb"].each { |f| require f }
 
-IRB.start(__FILE__)
+Pry.start
